+++ mex/common/models/base/schema.py
@@ -1,7 +1,10 @@
+from typing import Any
+
 from pydantic.json_schema import (
     GenerateJsonSchema as PydanticJsonSchemaGenerator,
 )
 from pydantic.json_schema import JsonSchemaValue
+from pydantic_core.core_schema import ComplexSchema
 
 
 class JsonSchemaGenerator(PydanticJsonSchemaGenerator):
@@ -18,3 +21,10 @@ class JsonSchemaGenerator(PydanticJsonSchemaGenerator):
         to stay compatible with mex-editor and mex-model.
         """
         return json_schema
+
+    def complex_schema(self, schema: ComplexSchema) -> dict[str, Any]:
+        """Returns a schema that matches a complex value."""
+        # TODO: clean this up in MX-1704 (stop-gap)
+        raise NotImplementedError(
+            "Method for generating JsonSchema for 'complex' schemas is not implemented."
+        )